tagaloop 2.0.0 makes IDs random, changes the format on disk.
I got sick of IDs which all look alike, so I changed the IDs to be random.
The effect of having string IDs is preserved by moving the timestamp into
the array that is stored for each individual item; [timestamp, label, string].

To convert from tagaloop 1.x to tagaloop 2.x, first, back up your file. (Tagaloop
will automatically do this by appending a ~ character, but you cannot be too
careful.) Then, open the file in tagaloop 1.x, and execute the following two
commands:

    - var a="ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789-.";function n(k){for(var t=0,i=0;i<k.length;i++)t=t*64+a.indexOf(k[i]);return t;}var newdata={};function newID(){function id(){return crypto.pseudoRandomBytes(6).toString('base64').replace(/\+/g,"-").replace(/\//g,".");}var o;do{o=id()}while(newdata.hasOwnProperty(o));return o;}keys(data).forEach(function(k){newdata[newID()]=[n(k)].concat(data[k]);});data=newdata;name='tagaloop 2.0.0';altered=true
    save exit

This should save a version of the file which is in the tagaloop 2.x format.